requirements.txt
streamlit
torch
diffusers
opencv-python
Pillow
numpy
facenet-pytorch
dlib
matplotlib
pyparsing
cycler
kiwisolver
imutils
codeformer-pip